Enum et::Mat

pub enum Mat<T> {}

Methods

impl<T> Mat<T>

fn from_elem((nrows, ncols): (u32, u32), elem: T) -> Box<Mat<T>> where T: Clone

Creates a matrix where each element is initialized to elem`elem`

fn from_fn<F>((nrows, ncols): (u32, u32), f: F) -> Box<Mat<T>> where F: FnMut((u32, u32)) -> T

Creates a matrix where each element is initialized using the function f`f`

fn reshape(slice: &[T], (nrows, ncols): (u32, u32)) -> &Mat<T>

Reshapes an slice into a matrix with dimensions (nrows, ncols)`(nrows, ncols)`

fn map<F>(&self, f: F) -> Map<F, &Mat<T>> where F: Fn<(T,)>, T: Clone

Lazily maps this matrix

fn ones((nrows, ncols): (u32, u32)) -> Box<Mat<T>> where T: Clone + One

Creates a matrix filled with ones

fn repr(&self) -> Mat<T>

Returns the raw representation of this matrix

fn zeros((nrows, ncols): (u32, u32)) -> Box<Mat<T>> where T: Clone + Zero

Creates a matrix filled with zeros